backend/requirements.txt
# Core Django
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-environ==0.11.2

# Cache & Queue
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-celery-beat==2.5.0
django-celery-results==2.2.0

# Authentication & Security
PyJWT==2.8.0
cryptography==42.0.2
pyotp==2.9.0
qrcode==7.4.2
django-ratelimit==4.1.0
django-csp==3.7
argon2-cffi==23.1.0

# API Documentation
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.1.1

# Validation & Serialization
marshmallow==3.20.2
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# File Handling
Pillow==10.2.0
python-magic==0.4.27

# Email & SMS
django-anymail==10.2
twilio==8.12.0

# Monitoring & Logging
prometheus-client==0.19.0
django-prometheus==2.3.1
structlog==23.2.0
sentry-sdk==1.40.6

# Development & Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0
responses==0.24.1

# Production
gunicorn==21.2.0
whitenoise==6.6.0
python-decouple==3.8

# Utilities
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0